# frontend/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BakeMate",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/shop/manage/configuration/": {
      "get": {
        "tags": [
          "Shop Management"
        ],
        "summary": "Get Shop Configuration",
        "description": "Retrieve the shop configuration for the authenticated baker.",
        "operationId": "get_shop_configuration_api_v1_shop_manage_configuration__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ShopConfigurationRead"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Shop Configuration Api V1 Shop Manage Configuration  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Shop Management"
        ],
        "summary": "Update Shop Configuration",
        "description": "Update the shop configuration for the authenticated baker.\nAssumes user has one config; if not, an ID would be needed.",
        "operationId": "update_shop_configuration_api_v1_shop_manage_configuration__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopConfigurationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopConfigurationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Shop Management"
        ],
        "summary": "Create Shop Configuration",
        "description": "Create a new shop configuration for the authenticated baker.\nA baker can only have one shop configuration.",
        "operationId": "create_shop_configuration_api_v1_shop_manage_configuration__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopConfigurationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopConfigurationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Shop Management"
        ],
        "summary": "Delete Shop Configuration",
        "description": "Delete the shop configuration for the authenticated baker.",
        "operationId": "delete_shop_configuration_api_v1_shop_manage_configuration__delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopConfigurationRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/shop/manage/configuration/embed-snippet/": {
      "get": {
        "tags": [
          "Shop Management"
        ],
        "summary": "Get Shop Embed Snippet",
        "description": "Get the embeddable HTML/JS snippet for the baker's shop.\nRequires an active shop configuration with a slug.",
        "operationId": "get_shop_embed_snippet_api_v1_shop_manage_configuration_embed_snippet__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Shop Embed Snippet Api V1 Shop Manage Configuration Embed Snippet  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/shop/public/{shop_slug}": {
      "get": {
        "tags": [
          "Public Shop"
        ],
        "summary": "View Public Shop",
        "description": "Retrieve the public view of a shop by its slug.\nOnly shows active shops that allow online orders.",
        "operationId": "view_public_shop_api_v1_shop_public__shop_slug__get",
        "parameters": [
          {
            "name": "shop_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicShopView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shop/public/{shop_slug}/order": {
      "post": {
        "tags": [
          "Public Shop"
        ],
        "summary": "Place Order From Public Shop",
        "description": "Place an order from a public shop.\nThe order will be created in the baker's main order system with status 'new-online'.\nConfirmation emails will be sent.",
        "operationId": "place_order_from_public_shop_api_v1_shop_public__shop_slug__order_post",
        "parameters": [
          {
            "name": "shop_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopOrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicShopProductView": {
        "properties": {
          "recipe_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recipe Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "recipe_id",
          "name",
          "price"
        ],
        "title": "PublicShopProductView"
      },
      "PublicShopView": {
        "properties": {
          "shop_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shop Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "theme_color_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Primary"
          },
          "theme_color_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Secondary"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/PublicShopProductView"
            },
            "type": "array",
            "title": "Products",
            "default": []
          },
          "delivery_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Options"
          },
          "min_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Amount"
          }
        },
        "type": "object",
        "title": "PublicShopView"
      },
...
      "ShopConfigurationCreate": {
        "properties": {
          "shop_slug": {
            "type": "string",
            "title": "Shop Slug"
          },
          "shop_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shop Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "theme_color_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Primary",
            "default": "#FFB6C1"
          },
          "theme_color_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Secondary",
            "default": "#F9F7F5"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShopStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": "inactive"
          },
          "allow_online_orders": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Online Orders",
            "default": false
          },
          "min_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Amount"
          },
          "max_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Order Amount"
          },
          "delivery_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Options"
          },
          "payment_methods_accepted": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Methods Accepted",
            "default": [
              "stripe"
            ]
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ShopProduct"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products",
            "default": []
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "shop_slug",
          "user_id"
        ],
        "title": "ShopConfigurationCreate"
      },
      "ShopConfigurationRead": {
        "properties": {
          "shop_slug": {
            "type": "string",
            "title": "Shop Slug"
          },
          "shop_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shop Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "theme_color_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Primary",
            "default": "#FFB6C1"
          },
          "theme_color_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Secondary",
            "default": "#F9F7F5"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShopStatus"
              },
              {
                "type": "null"
              }
            ],
            "default": "inactive"
          },
          "allow_online_orders": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Online Orders",
            "default": false
          },
          "min_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Amount"
          },
          "max_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Order Amount"
          },
          "delivery_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Options"
          },
          "payment_methods_accepted": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Methods Accepted",
            "default": [
              "stripe"
            ]
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ShopProduct"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products",
            "default": []
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "shop_slug",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ShopConfigurationRead"
      },
      "ShopConfigurationUpdate": {
        "properties": {
          "shop_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shop Slug"
          },
          "shop_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shop Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "theme_color_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Primary"
          },
          "theme_color_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color Secondary"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShopStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "allow_online_orders": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Online Orders"
          },
          "min_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Order Amount"
          },
          "max_order_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Order Amount"
          },
          "delivery_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Options"
          },
          "payment_methods_accepted": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Methods Accepted"
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ShopProduct"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products"
          }
        },
        "type": "object",
        "title": "ShopConfigurationUpdate"
      },
      "ShopOrderCreate": {
        "properties": {
          "customer_name": {
            "type": "string",
            "title": "Customer Name"
          },
          "customer_email": {
            "type": "string",
            "title": "Customer Email"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "delivery_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Address"
          },
          "delivery_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Instructions"
          },
          "pickup_time_slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Time Slot"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShopOrderItemCreate"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "shop_slug": {
            "type": "string",
            "title": "Shop Slug"
          }
        },
        "type": "object",
        "required": [
          "customer_name",
          "customer_email",
          "shop_slug"
        ],
        "title": "ShopOrderCreate"
      },
...
      "ShopProduct": {
        "properties": {
          "recipe_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recipe Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "recipe_id",
          "name",
          "price"
        ],
        "title": "ShopProduct"
      },
      "ShopStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "maintenance"
        ],
        "title": "ShopStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login/access-token"
          }
        }
      }
    }
  }
}